// web/src/components/clouds/pieces.generated.ts
// Generated by assets-src/clouds/generate-clouds.mjs. Do not edit by hand.

export interface CloudPiece {
  name: string;
  side: "left" | "right";
  anchor: "top" | "bottom";
  /** Distance from the anchor edge, in % of the hero height. */
  offset: number;
  /** Height in % of the hero height. */
  height: number;
  width: number;
  pixelHeight: number;
}

export const CLOUD_PIECES: CloudPiece[] = [
  {
    "name": "left-top",
    "side": "left",
    "anchor": "top",
    "offset": 0,
    "height": 43.333,
    "width": 318,
    "pixelHeight": 390
  },
  {
    "name": "left-main",
    "side": "left",
    "anchor": "bottom",
    "offset": -4,
    "height": 76.667,
    "width": 654,
    "pixelHeight": 690
  },
  {
    "name": "right-top",
    "side": "right",
    "anchor": "top",
    "offset": 7.333,
    "height": 68.667,
    "width": 410,
    "pixelHeight": 618
  },
  {
    "name": "right-bottom",
    "side": "right",
    "anchor": "bottom",
    "offset": -4,
    "height": 50,
    "width": 440,
    "pixelHeight": 450
  }
];